% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seven31.R
\name{compare}
\alias{compare}
\title{compare the bits of two double}
\usage{
compare(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{another number}
}
\description{
compare the bits of two double
}
\examples{
compare( .3 + .6, .9)
compare( 2, sqrt(2)^2)

}
