% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seven31.R
\name{reveal}
\alias{reveal}
\title{reveal the bits of a double}
\usage{
reveal(x)
}
\arguments{
\item{x}{a double}
}
\description{
reveal the bits of a double
}
\examples{
reveal( 1.0 )
reveal( 0.0 )
reveal( -0.0 )
reveal( NA )
reveal( NaN )

}
