% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seven31.R
\name{reveal}
\alias{reveal}
\title{reveal the bits of a double}
\usage{
reveal(...)
}
\arguments{
\item{\dots}{several expression that should each evaluate to a single double}
}
\description{
reveal the bits of a double
}
\examples{
reveal( 1.0, 0.0, -0.0 )
reveal( NA, NaN )

}
